Setup basic game functionality
This piece of code sets up the basic game functionality.
Initialized the guess_word variable to store the randomly selected word. Blanks variable stores the string of blanks according to the length of word. guessed_letter stores the letter input by user. While loop is initialized to ask for user input continuously. if the input letter is in the word, it will print "Right" otherwise prints "wrong".