Created main caesarCipher( ) function and added condition for numeric
Removed the separate encrypt and decrypt function and created single function CaesarCipher() to contain all code and run this function only to run the program.

Added conditions that if the element of the input_text is numeric then it is added in the Cipher string with no changes while encoding and decoding.